+++ ion-ripple.js
@@ -4,8 +4,6 @@
   angular.module('ionMDRipple', ['ionic'])
     .directive('ionRipple', ionRipple);
 
-  ionRipple.$inject = [];
-
   function ionRipple() {
     var directive = {
       restrict: 'A',
@@ -26,55 +24,33 @@
       ripple.classList.add(rippleClass);
       element[0].insertBefore(ripple, element[0].firstChild);
 
-      element.on('touchend mouseup', rippleHandler);
+      element.on('click', rippleHandler);
+      angular.element(ripple).on('animationend webkitAnimationEnd', deactivateRipple);
 
       //remove the event listener on scope destroy
       scope.$on('$destroy', function() {
-        element.off('touchend mouseup', rippleHandler);
+        element.off('click', rippleHandler);
+        angular.element(ripple).off('animationend webkitAnimationEnd', deactivateRipple);
       });
 
-      function rippleHandler(event) {
-        ripple = element[0].querySelector('.' + rippleClass);
-
-        // Remove animation effect
+      function deactivateRipple() {
         ripple.classList.remove(animateClassName);
+      }
 
-        // Set ripple size
+      function rippleHandler(event) {
         if (!ripple.offsetHeight && !ripple.offsetWidth) {
           size = Math.max(element[0].offsetWidth, element[0].offsetHeight);
           ripple.style.width = ripple.style.height = size + 'px';
         }
-        switch (event.type) {
-          case 'mouseup':
-            x = event.pageX;
-            y = event.pageY;
-            break;
-          case 'touchend':
-            try {
-              var origEvent;
 
-              if (typeof event.changedTouches !== 'undefined') {
-                origEvent = event.changedTouches[0];
-              } else {
-                origEvent = event.originalEvent;
-              }
+        x = event.pageX;
+        y = event.pageY;
 
-              x = origEvent.pageX;
-              y = origEvent.pageY;
-            } catch (e) {
-              // fall back to center of el
-              x = ripple.offsetWidth / 2;
-              y = ripple.offsetHeight / 2;
-            }
-            break;
-        }
-
-        // set new ripple position by click or touch position
         function getPos(element) {
           var de = document.documentElement;
           var box = element.getBoundingClientRect();
-          var top = box.top + window.pageYOffset - de.clientTop;
-          var left = box.left + window.pageXOffset - de.clientLeft;
+          var top = box.top + pageYOffset - de.clientTop;
+          var left = box.left + pageXOffset - de.clientLeft;
 
           return { top: top, left: left };
         }
@@ -83,7 +59,6 @@
         ripple.style.top = (y - offsets.top - size / 2) + 'px';
         ripple.style.left = (x - offsets.left - size / 2) + 'px';
 
-        // Add animation effect
         ripple.classList.add(animateClassName);
       }
     }